+++ tests/test_format_strings.py
@@ -7,5 +7,5 @@ def test_old_format_string(v):
 
 
 def test_new_format_string(v):
-    v.scan("'{a}, {b:0d}'.format(**locals())")
-    check(v.used_names, ['a', 'b', 'locals'])
+    v.scan("'{a}, {b:0d} {c:<30} {d:.2%}'.format(**locals())")
+    check(v.used_names, ['a', 'b', 'c', 'd', 'locals'])